Extract shader object
